+++ src/humlib.cpp
@@ -1,7 +1,7 @@
 //
 // Programmer:    Craig Stuart Sapp <craig@ccrma.stanford.edu>
 // Creation Date: Sat Aug  8 12:24:49 PDT 2015
-// Last Modified: Sun Jul 14 15:38:39 CEST 2019
+// Last Modified: Sun Jul 14 15:59:13 CEST 2019
 // Filename:      /include/humlib.cpp
 // URL:           https://github.com/craigsapp/humlib/blob/master/src/humlib.cpp
 // Syntax:        C++11
@@ -19329,7 +19329,7 @@ HumNum HumdrumFileStructure::getBarlineDurationToEnd(int index) const {
 //
 
 void HumdrumFileStructure::setLineRhythmAnalyzed(void) {
-	for (int i=0; i<m_lines.size(); i++) {
+	for (int i=0; i<(int)m_lines.size(); i++) {
 		m_lines[i]->m_rhythm_analyzed = true;
 	}
 }
